refactor: move controllers to similar files
