+++ controllers/contacts/index.js
@@ -1,17 +1,15 @@
-const {
-    listContacts,
-    getContactById,
-    removeContact,
-    addContact,
-    updateContact,
-    updateStatusContact,
-  } = require('./contacts');
+const getContactById = require("./getContactById");
+const listContacts = require("./listContacts");
+const addContact = require("./addContact");
+const removeContact = require("./removeContact");
+const updateContact = require("./updateContact");
+const updateStatusContact = require("./updateStatusContact");
 
-  module.exports = {
-    listContacts,
-    getContactById,
-    removeContact,
-    addContact,
-    updateContact,
-    updateStatusContact
-  };
+module.exports = {
+  listContacts,
+  getContactById,
+  removeContact,
+  addContact,
+  updateContact,
+  updateStatusContact,
+};
