+++ controllers/contacts/removeContact.js
@@ -1,0 +1,9 @@
+const Contact = require("../../models/contact");
+
+async function removeContact(contactId) {
+    const contactToDelete = await Contact.findByIdAndRemove({ _id: contactId });
+    return contactToDelete;
+  };
+
+
+module.exports = removeContact;
