+++ controllers/contacts/updateStatusContact.js
@@ -1,0 +1,16 @@
+const Contact = require("../../models/contact");
+
+
+
+
+
+async function updateStatusContact(contactId, body) {
+  const contactToUpdate = await Contact.findOneAndUpdate(
+    { _id: contactId },
+    { $set: body },
+    { new: true }
+  );
+  return contactToUpdate;
+}
+
+module.exports = updateStatusContact;
